some modal corrections and requests

+++ frontend/src/components/modals/category/ModalEditCategory.jsx
@@ -3,12 +3,11 @@ import { Modal, Button, Form, Card } from 'react-bootstrap';
 import { useTranslation } from 'react-i18next';
 import axios from 'axios';
 import { useState } from 'react';
-import { useAuth } from '../../context/AuthContext'
+import { useAuth } from '../../context/AuthContext';
 
-const ModalEditCategory = ({ showModal, onClose, selectedCategoryUUID }) => {
-  const [formData, setFormData] = useState([]);
+const ModalEditCategory = ({ showModal, onClose, selectedCategoryUUID, handleCategoryChange }) => {
+  const [formData, setFormData] = useState({});
   const { getTokenHeader } = useAuth();
-  // console.log(selectedCategoryUUID);
 
   useEffect(() => {
     const fetchCategoryData = async () => {
@@ -19,16 +18,15 @@ const ModalEditCategory = ({ showModal, onClose, selectedCategoryUUID }) => {
             headers: getTokenHeader()
           }
         );
-
         const categoryData = response.data;
         setFormData((prevFormData) => ({
           ...prevFormData,
-          category_name_lt: categoryData.data.nameLt || '',
-          category_name_en: categoryData.data.nameEn || '',
-          category_description_lt: categoryData.data.descriptionLt || '',
-          category_description_en: categoryData.data.descriptionEn || '',
-          album_type: categoryData.data.albumType || '',
-          photo_limit: categoryData.data.photoLimit || ''
+          category_nameLt: categoryData.nameLt || '',
+          category_nameEn: categoryData.nameEn || '',
+          category_descriptionLt: categoryData.descriptionLt || '',
+          category_descriptionEn: categoryData.descriptionEn || '',
+          albumType: categoryData.albumType || '',
+          photoLimit: categoryData.photoLimit || ''
         }));
       } catch (error) {
         console.log('Error fetching category data:', error);
@@ -38,48 +36,35 @@ const ModalEditCategory = ({ showModal, onClose, selectedCategoryUUID }) => {
     if (selectedCategoryUUID) {
       fetchCategoryData();
     }
-  }, [selectedCategoryUUID, getTokenHeader, setFormData]);
+  }, [selectedCategoryUUID, getTokenHeader]);
 
-  const confirmSave = async () => {
+  const handleCategoryAction = async (action) => {
     try {
-      await axios.put(`http://localhost:8080/api/v1/category/${selectedCategoryUUID}`, formData, {
-        headers: getTokenHeader()
-      });
+      if (action === 'delete') {
+        await axios.delete(`http://localhost:8080/api/v1/category/${selectedCategoryUUID}`, {
+          headers: getTokenHeader()
+        });
+        // Update parent component state to remove the deleted category
+        handleCategoryChange(selectedCategoryUUID);
+      } else if (action === 'update') {
+        await axios.put(`http://localhost:8080/api/v1/category/${selectedCategoryUUID}`, formData, {
+          headers: getTokenHeader()
+        });
+        // Update parent component state to reflect changes in the updated category
+        handleCategoryChange(selectedCategoryUUID, formData);
+      }
     } catch (error) {
-      console.error('Error saving competition:', error);
+      console.error(`Error ${action === 'delete' ? 'deleting' : 'updating'} category:`, error);
     }
-    onClose(true);
-  };
-
-  const handleDelete = async () => {
-    try {
-      await axios.delete(`http://localhost:8080/api/v1/category/${selectedCategoryUUID}`, {
-        headers: getTokenHeader()
-      });
-    } catch (error) {
-      console.log('Error deleting competition', error);
-    }
-    onClose(true);
+    onClose();
   };
 
   const handleInputChange2 = (event) => {
     const { name, value } = event.target;
-    if (name === 'photo_limit') {
-      if (value === '' || (parseInt(value) >= 1 && parseInt(value) <= 15)) {
-        setFormData({
-          ...formData,
-          [name]: value
-        });
-        setPhotoLimitError('');
-      } else {
-        setPhotoLimitError(t('editcomp.limitError'));
-      }
-    } else {
-      setFormData({
-        ...formData,
-        [name]: value
-      });
-    }
+    setFormData({
+      ...formData,
+      [name]: value
+    });
   };
 
   const [t] = useTranslation();
@@ -92,42 +77,42 @@ const ModalEditCategory = ({ showModal, onClose, selectedCategoryUUID }) => {
         <Form.Label> {t('modalCategory.titleCat_en')}</Form.Label>
         <Form.Control
           type="text"
-          name="category_name_en"
-          id="category_name_en"
-          value={formData.category_name_en || ''}
+          name="category_nameEn"
+          id="category_nameEn"
+          value={formData.category_nameEn || ''}
           onChange={handleInputChange2}
         ></Form.Control>
         <Form.Label> {t('modalCategory.titleCat_lt')}</Form.Label>
         <Form.Control
           type="text"
-          name="category_name_lt"
-          id="category_name_lt"
-          value={formData.category_name_lt || ''}
+          name="category_nameLt"
+          id="category_nameLt"
+          value={formData.category_nameLt || ''}
           onChange={handleInputChange2}
         ></Form.Control>
         <Form.Label> {t('modalCategory.description_lt')}</Form.Label>
         <Form.Control
           type="text"
-          name="category_description_lt"
-          id="category_description_lt"
-          value={formData.category_description_lt || ''}
+          name="category_descriptionLt"
+          id="category_descriptionLt"
+          value={formData.category_descriptionLt || ''}
           onChange={handleInputChange2}
         ></Form.Control>
         <Form.Label> {t('modalCategory.description_en')}</Form.Label>
         <Form.Control
           type="text"
-          name="category_description_en"
-          id="category_description_en"
-          value={formData.category_description_en || ''}
+          name="category_descriptionEn"
+          id="category_descriptionEn"
+          value={formData.category_descriptionEn || ''}
           onChange={handleInputChange2}
         ></Form.Control>
         <Form.Label> {t('modalCategory.type')}</Form.Label>
         <Form.Control
           as="select"
-          name="album_type"
-          id="album_type"
+          name="albumType"
+          id="albumType"
           onChange={handleInputChange2}
-          value={formData.album_type || ''}
+          value={formData.albumType || ''}
         >
           <option value=""></option>
           <option value="SINGLE">{t('modalCategory.single')}</option>
@@ -135,9 +120,9 @@ const ModalEditCategory = ({ showModal, onClose, selectedCategoryUUID }) => {
         </Form.Control>
         <Form.Label>{t('modalCategory.photoLimit')}</Form.Label>
         <Form.Control
-          name="photo_limit"
-          id="photo_limit"
-          value={formData.photo_limit || ''}
+          name="photoLimit"
+          id="photoLimit"
+          value={formData.photoLimit || ''}
           onChange={handleInputChange2}
           min="1"
           max="15"
@@ -147,10 +132,10 @@ const ModalEditCategory = ({ showModal, onClose, selectedCategoryUUID }) => {
         <Button variant="secondary" onClick={onClose}>
           {t('modalCategory.closeButton')}
         </Button>
-        <Button variant="secondary" onClick={confirmSave}>
+        <Button variant="secondary" onClick={() => handleCategoryAction('update')}>
           {t('modalCategory.updateCategory')}
         </Button>
-        <Button variant="secondary" onClick={handleDelete}>
+        <Button variant="secondary" onClick={() => handleCategoryAction('delete')}>
           {t('modalCategory.deleteCategory')}
         </Button>
       </Modal.Footer>
